backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "Backend for the Todo app with Cohere Chatbot Integration"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.115.0"
uvicorn = {extras = ["standard"], version = "0.32.0"}
sqlmodel = "0.0.22"
pydantic = "2.10.1"
pydantic-settings = "2.7.0"
python-multipart = "0.0.20"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
alembic = "1.14.0"
asyncpg = "0.31.0"
sqlalchemy = {extras = ["asyncio"], version = "2.0.23"}
cohere = "5.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"